// src/models/mutate-user-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * proxima-nexus-data-plane-api
 * Proxima Nexus Data Plane API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface MutateUserResponseDto {
    /**
     * User ID
     */
    'userId': string;
}