// src/models/mutate-user-connection-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * proxima-nexus-data-plane-api
 * Proxima Nexus Data Plane API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface MutateUserConnectionDto {
    /**
     * Type of connection to create or modify
     */
    'type': MutateUserConnectionDtoTypeEnum;
}

export enum MutateUserConnectionDtoTypeEnum {
    friend = 'friend',
    blocked = 'blocked'
}